% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{code_covergae}
\alias{code_covergae}
\title{Calculate slicing coverage for code directly}
\usage{
code_covergae(source_files, test_files)
}
\arguments{
\item{source_files}{Character vector of source files with function definitions to measure coverage}

\item{test_files}{Character vector of test files with code to test the functions}
}
\description{
Calculate slicing coverage for code directly
}
