% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{file_coverage}
\alias{file_coverage}
\title{Calculate slicing coverage for a set of files}
\usage{
file_coverage(
  source_files,
  test_files,
  line_exclusions = NULL,
  function_exclusions = NULL
)
}
\arguments{
\item{source_files}{Character vector of source files with function definitions to measure coverage}

\item{test_files}{Character vector of test files with code to test the functions}

\item{line_exclusions}{Currently unsupported}

\item{function_exclusions}{Currently unsupported}
}
\description{
Calculate slicing coverage for a set of files
}
