% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{maximum_coverage}
\alias{maximum_coverage}
\title{Calculate the maximum possible slicing coverage with the current assertions.
Considering this value can indicate whether the slicing coverage score is limited by
the number of assertions or the amount of covered code.}
\usage{
maximum_coverage(f, ...)
}
\arguments{
\item{f}{Function to calculate slicing coverage}

\item{...}{Arguments to pass to the function}
}
\description{
Calculate the maximum possible slicing coverage with the current assertions.
Considering this value can indicate whether the slicing coverage score is limited by
the number of assertions or the amount of covered code.
}
