% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{package_coverage}
\alias{package_coverage}
\title{Calculate slicing coverage for a package}
\usage{
package_coverage(path = ".")
}
\arguments{
\item{path}{Path to the package}
}
\description{
Calculate slicing coverage for a package
}
