% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{with_options}
\alias{with_options}
\title{Executes an expression with the given options set and resets them afterwards.}
\usage{
with_options(options, expr)
}
\arguments{
\item{options}{A list of options to set.}

\item{expr}{The expression to execute}
}
\value{
The result of the expression.
}
\description{
Executes an expression with the given options set and resets them afterwards.
}
