% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{configure}
\alias{configure}
\title{Configures various options for this package. If no value is given for any
option, it's default is used.}
\usage{
configure(
  flowr_host = NULL,
  flowr_port = NULL,
  measure_time = NULL,
  return_covr_result = NULL,
  slicing_points = NULL,
  log_level = NULL,
  unknown_locations = NULL
)
}
\arguments{
\item{flowr_host}{The host where the flowr server is running. Defaults to
"localhost".}

\item{flowr_port}{The port where the flowr server is running. Defaults to
1042.}

\item{measure_time}{Whether to measure the execution time of different
calculations. Defaults to FALSE.}

\item{return_covr_result}{Whether to also return the coverage covr calculated}

\item{slicing_points}{Whether to return all slicing points we found and slices for}

\item{log_level}{The minimum level of logs that should be shown}

\item{unknown_locations}{Whether to return the number of source refs flowr did not
know about}
}
\description{
Configures various options for this package. If no value is given for any
option, it's default is used.
}
